requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
pyirtoy==0.2.6
pyserial==2.7
requests==2.7.0
uWSGI==2.0.10
Werkzeug==0.10.4
